resnet-50/requirements.txt
datasets==3.1.0
Flask==2.2.5
Pillow==11.0.0
Requests==2.32.3
safetensors==0.4.5
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3
